+++ pkg/app/app.go
@@ -55,8 +55,8 @@ func New() EnvdApp {
 		CommandBuild,
 		CommandContext,
 		CommandDestroy,
-		CommandGetEnvironment,
-		CommandGetImage,
+		CommandEnvironment,
+		CommandImage,
 		CommandInit,
 		CommandPause,
 		CommandPrune,
